Share value scratch for low register masks
Low register-part definedness is a transient module-scope verification fact, so it should not allocate a second value-id table every time the low verifier runs. Replace the ordinal-only module scratch with an exclusive u32 scratch table that can be acquired either as local value ordinals or as zeroed verifier payloads. The low verifier now takes scratch explicitly, keeps the module const, zeros the active value range once per module, and amortizes symbol facts while resetting only per-function walk storage.

Register-part masks are narrowed to u32 across descriptor data, generated C, and diagnostics. That matches the shared scratch width and keeps partial-register modelling compact while still representing the current target descriptor cases.

AMDGPU HAL-native preflight is materialized after the stable final emission frame is returned. The preflight facts retain schedule and allocation pointers, so running it through the frame-builder validation callback captured temporary frame addresses and made later HSACO record construction reject otherwise valid kernels.

+++ loom/src/loom/codegen/low/descriptors.h
@@ -117,7 +117,7 @@ typedef uint16_t loom_low_reg_class_alt_flags_t;
 typedef uint16_t loom_low_reg_class_flags_t;
 
 // Register-part definedness mask within one allocation unit.
-typedef uint64_t loom_low_register_part_mask_t;
+typedef uint32_t loom_low_register_part_mask_t;
 
 // Register class is virtual-only and has no physical register inventory.
 #define LOOM_LOW_REG_CLASS_FLAG_VIRTUAL_ONLY ((uint16_t)1u << 0)
